+++ src/components/film-card.js
@@ -1,7 +1,7 @@
 export const getFilmCardTemplate = (film) => {
   const {
     title,
-    poster,
+    posterUrl,
     description,
     rating,
     releaseDate,
@@ -21,7 +21,7 @@ export const getFilmCardTemplate = (film) => {
           <span class="film-card__duration">${duration}</span>
           <span class="film-card__genre">${genres}</span>
         </p>
-        <img src="${poster}" alt="" class="film-card__poster">
+        <img src="${posterUrl}" alt="" class="film-card__poster">
         <p class="film-card__description">${description}</p>
         <a class="film-card__comments">${comments.length} comments</a>
         <form class="film-card__controls">
